# openapi/family-revision.yaml
openapi: 3.0.3
info:
  title: API для работы с данными по переписям (РС/ИВ/ПП)
  version: "1.0.1"
...
paths:
...
  /family-revision/family:
    post:
      tags:
        - FamilyRevision
      operationId: getFamilyRevisionByNum
      summary: "Family revision filter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "./model/genealogy/family.revision/family-revision-filter.yaml"
        required: true
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyRevisionResponse"
        404:
          description: "Failed to find data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyRevisionErrorResponse"
components:
  schemas:
    FamilyRevisionResponse:
      type: object
      oneOf:
        - $ref: "./model/genealogy/family.revision/family-revision.yaml"
        - $ref: "#/components/schemas/FamilyRevisionErrorResponse"
        - $ref: "#/components/schemas/ArchiveWithFamilyRevisionList"
    ArchiveWithFamilyRevisionList:
      type: object
      description: Список архивов с вложенным списком ревизий
      properties:
        data:
          description: Список архивов с вложенным списком ревизий
          type: array
          items:
            $ref: "./model/genealogy/family.revision/archive-with-family-revision.yaml"
    FamilyRevisionErrorResponse:
      $ref: "./model/common/error-response.yaml"